parge register from cpu
